requirements.txt
streamlit
pandas
numpy
requests
python-dotenv
openai
streamlit-autorefresh
